+++ src/include/rest_catalog/objects/blob_metadata.hpp
@@ -30,7 +30,7 @@ public:
 	string type;
 	int64_t snapshot_id;
 	int64_t sequence_number;
-	vector<int64_t> fields;
+	vector<int32_t> fields;
 	case_insensitive_map_t<string> properties;
 	bool has_properties = false;
 };
