+++ src/include/rest_catalog/objects/set_current_view_version_update.hpp
@@ -29,7 +29,7 @@ public:
 
 public:
 	BaseUpdate base_update;
-	int64_t view_version_id;
+	int32_t view_version_id;
 	string action;
 	bool has_action = false;
 };
