+++ src/include/rest_catalog/objects/set_default_sort_order_update.hpp
@@ -29,7 +29,7 @@ public:
 
 public:
 	BaseUpdate base_update;
-	int64_t sort_order_id;
+	int32_t sort_order_id;
 	string action;
 	bool has_action = false;
 };
