+++ src/rest_catalog/objects/equality_delete_file.cpp
@@ -50,7 +50,7 @@ string EqualityDeleteFile::TryFromJSON(yyjson_val *obj) {
 			size_t idx, max;
 			yyjson_val *val;
 			yyjson_arr_foreach(equality_ids_val, idx, max, val) {
-				int64_t tmp;
+				int32_t tmp;
 				if (yyjson_is_int(val)) {
 					tmp = yyjson_get_int(val);
 				} else {
