+++ src/rest_catalog/objects/remove_partition_specs_update.cpp
@@ -39,7 +39,7 @@ string RemovePartitionSpecsUpdate::TryFromJSON(yyjson_val *obj) {
 			size_t idx, max;
 			yyjson_val *val;
 			yyjson_arr_foreach(spec_ids_val, idx, max, val) {
-				int64_t tmp;
+				int32_t tmp;
 				if (yyjson_is_int(val)) {
 					tmp = yyjson_get_int(val);
 				} else {
