+++ src/rest_catalog/objects/schema.cpp
@@ -46,7 +46,7 @@ string Schema::Object1::TryFromJSON(yyjson_val *obj) {
 			size_t idx, max;
 			yyjson_val *val;
 			yyjson_arr_foreach(identifier_field_ids_val, idx, max, val) {
-				int64_t tmp;
+				int32_t tmp;
 				if (yyjson_is_int(val)) {
 					tmp = yyjson_get_int(val);
 				} else {
